Implement undo history for `cp`, `mv` and `rm` and `DEBUG` mode for stack trace

+++ src/commands/core/file_operations.py
@@ -1,11 +1,55 @@
 import os
 import shutil
+import shlex
 
 from argparse import ArgumentParser
 from src.command import command, CommandEnv
 from src.path import validate_path
 
-from src.constants import ROOT_DIR
+from src.constants import TRASH_DIR
+from src.constants import UNDO_HISTORY_PATH
+
+
+def get_next_undo_count() -> int:
+    if not os.path.exists(UNDO_HISTORY_PATH):
+        return 1
+
+    with open(UNDO_HISTORY_PATH, 'r') as f:
+        lines = f.readlines()
+
+    return len(lines) + 1
+
+
+def get_undo_prefix() -> str:
+    return f"{get_next_undo_count()}_"
+
+
+def add_undo_command(name: str, args: list[str]) -> None:
+    args.insert(0, name)
+
+    with open(UNDO_HISTORY_PATH, 'a') as f:
+        f.write(f"{shlex.join(args)}\n")
+
+
+def copy_and_trash_overwritten_function(undo_prefix: str, dest_path: str, undo_paths: list[str], add_all_paths: bool):
+    def _copy_and_trash_overwritten(src: str, dst: str, *, follow_symlinks=True):
+        dst_exists = os.path.exists(dst)
+
+        if dst_exists:
+            rel_ovewritten_trash_path = os.path.relpath(dst, dest_path)
+            ovewritten_trash_dir = undo_prefix + os.path.basename(dest_path)
+            ovewritten_trash_path = os.path.join(TRASH_DIR, ovewritten_trash_dir, rel_ovewritten_trash_path)
+
+            os.makedirs(os.path.dirname(ovewritten_trash_path), exist_ok=True)
+            shutil.copy2(dst, ovewritten_trash_path)
+
+        if not dst_exists or add_all_paths:
+            undo_paths.append(dst)
+
+        shutil.copy2(src, dst, follow_symlinks=follow_symlinks)
+
+    return _copy_and_trash_overwritten
+
 
 @command(
     name="cp",
@@ -28,29 +72,44 @@ def cmd_cp(env: CommandEnv, args: list[str]) -> None:
     validate_path(source_path)
 
     dest_path = env.get_path(argv.dest)
-
-    log_dest_path = dest_path
     if os.path.isdir(dest_path):
-        log_dest_path = os.path.join(dest_path, os.path.basename(source_path))
+        dest_path = os.path.join(dest_path, os.path.basename(source_path))
 
+    undo_prefix = get_undo_prefix()
+    undo_paths: list[str] = [dest_path]
+
+    copy_and_trash_overwritten = copy_and_trash_overwritten_function(
+        undo_prefix=undo_prefix,
+        dest_path=dest_path,
+        undo_paths=undo_paths,
+        add_all_paths=False  # we don't need to remove files that are going to be overwritten anyway during undo
+    )
+
+    # TODO: maybe use copy_and_trash_overwritten for single file copies aswell?
     if os.path.isfile(source_path):
         try:
+            if os.path.isfile(dest_path):
+                ovewritten_trash_name = undo_prefix + os.path.basename(dest_path)
+                ovewritten_trash_path = os.path.join(TRASH_DIR, ovewritten_trash_name)
+
+                shutil.copy2(dest_path, ovewritten_trash_path)
+
             shutil.copy2(source_path, dest_path)
         except PermissionError:
             raise PermissionError("No permission")
 
-    elif os.path.isdir(source_path):
+    if os.path.isdir(source_path):
         if not argv.r:
             raise IsADirectoryError("Unable to copy directory without '-r' flag present")
 
         try:
-            shutil.copytree(source_path, dest_path)
+            shutil.copytree(source_path, dest_path, dirs_exist_ok=True, copy_function=copy_and_trash_overwritten)
         except PermissionError:
             raise PermissionError("No permission")
-        except FileExistsError:
-            raise FileExistsError(f"Destination is a file {dest_path}")
 
-    env.log_success(f"Successfully copied {source_path} to {log_dest_path}")
+    add_undo_command('cp', undo_paths)
+
+    env.log_success(f"Successfully copied {source_path} to {dest_path}")
 
 
 @command(
@@ -71,19 +130,44 @@ def cmd_mv(env: CommandEnv, args: list[str]) -> None:
     validate_path(source_path)
 
     dest_path = env.get_path(argv.dest)
-
-    log_dest_path = dest_path
     if os.path.isdir(dest_path):
-        log_dest_path = os.path.join(dest_path, os.path.basename(source_path))
+        dest_path = os.path.join(dest_path, os.path.basename(source_path))
 
-    try:
-        shutil.move(source_path, dest_path)
-    except PermissionError:
-        raise PermissionError("No permission")
-    except FileExistsError:
-        raise FileExistsError(f"Destination is a file {log_dest_path}")
+    undo_prefix = get_undo_prefix()
+    undo_paths: list[str] = [source_path, dest_path]
 
-    env.log_success(f"Successfully moved {source_path} to {log_dest_path}")
+    copy_and_trash_overwritten = copy_and_trash_overwritten_function(
+        undo_prefix=undo_prefix,
+        dest_path=dest_path,
+        undo_paths=undo_paths,
+        add_all_paths=True  # we need to move all files back to where they were during undo
+    )
+
+    if os.path.isfile(source_path):
+        try:
+            if os.path.isfile(dest_path):
+                ovewritten_trash_name = undo_prefix + os.path.basename(dest_path)
+                ovewritten_trash_path = os.path.join(TRASH_DIR, ovewritten_trash_name)
+
+                shutil.copy2(dest_path, ovewritten_trash_path)
+
+            shutil.move(source_path, dest_path)
+        except PermissionError:
+            raise PermissionError("No permission")
+        except FileExistsError:
+            raise FileExistsError(f"Destination is a file {dest_path}")
+
+    if os.path.isdir(source_path):
+        try:
+            shutil.move(source_path, dest_path, copy_and_trash_overwritten)
+        except PermissionError:
+            raise PermissionError("No permission")
+        except FileExistsError:
+            raise FileExistsError(f"Destination is a file {dest_path}")
+
+    add_undo_command('mv', undo_paths)
+
+    env.log_success(f"Successfully moved {source_path} to {dest_path}")
 
 
 @command(
@@ -122,20 +206,22 @@ def cmd_rm(env: CommandEnv, args: list[str]) -> None:
             env.log_success("User did not confirm prompt. Done nothing")
             return
 
-    trash_path = os.path.join(ROOT_DIR, ".trash")
-    if not os.path.exists(trash_path):
-        os.mkdir(trash_path)
+    if not os.path.exists(TRASH_DIR):
+        os.mkdir(TRASH_DIR)
 
-    log_dest_path = os.path.join(trash_path, os.path.basename(source_path))
+    undo_prefix = get_undo_prefix()
+
+    trashed_source_name = undo_prefix + os.path.basename(source_path)
+    trashed_source_path = os.path.join(TRASH_DIR, trashed_source_name)
 
     if not argv.test:
         try:
-            shutil.move(source_path, trash_path)
+            shutil.move(source_path, trashed_source_path)
         except PermissionError:
             raise PermissionError("No permission")
-        except FileExistsError:
-            raise FileExistsError(f"Destination is a file {log_dest_path}")
 
-        env.log_success(f"Successfully moved {source_path} to {log_dest_path}")
+        env.log_success(f"Successfully moved {source_path} to {trashed_source_path}")
     else:
         env.log_success("Test flag present. Done nothing")
+
+    add_undo_command('rm', [source_path])
